Updates in several classes
- Logging: changes in the color scheme.
- ILObfuscator: verification of both "division" and "modulo" operations
in each function. They impact the usage of original variables in fake
instructions.
- InternalMethods: correction of bug in UpdateAllCalls; new method to
check whether a function has "division" and/or "modulo" operations.
- InternalTypes: new attribute to identify the functions that have
"division" and/or "modulo" operations
- Meshing: change in the name of parameters; new conditions to decide
whether a basic block will involve either "FakeVariablesOnly" or "Both".
- ExportImport: comments in the method used for CommonModule.
- Assembler: support any number of functions in a routine. The last one
is considered the "main".
- PseudoCode: correction of minor bugs; improvement in the support for
embedded (if/loop) instructions; creation of a separated basic block for
the "return" instruction in case the function contains "division" and/or
"modulo" instructions. It's needed to be able to provided a wrong answer
with invalid parameters without getting an runtime error.
- New files: test cases